requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
ansicon==1.89.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asteroid-filterbanks==0.4.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.0
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.7
croniter==1.3.14
cycler==0.11.0
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
docopt==0.6.2
einops==0.6.1
fastapi==0.88.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
greenlet==2.0.2
h11==0.14.0
huggingface-hub==0.14.1
HyperPyYAML==1.2.0
idna==3.4
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinxed==1.2.0
joblib==1.2.0
julius==0.2.7
kiwisolver==1.4.4
lazy_loader==0.2
librosa==0.10.0.post2
lightning==2.0.2
lightning-cloud==0.5.34
lightning-utilities==0.8.0
llvmlite==0.39.1
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
networkx==3.1
numba==0.56.4
numpy==1.23.5
omegaconf==2.3.0
openai==0.27.4
optuna==3.1.1
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
parse==1.19.0
Pillow==9.5.0
pooch==1.6.0
primePy==1.3
protobuf==3.20.3
psutil==5.9.5
pyannote.audio==2.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pysrt==1.1.2
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytorch-metric-learning==2.1.0
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
semver==3.0.0
sentencepiece==0.1.98
shellingham==1.5.0.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.14
SQLAlchemy==2.0.10
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboardX==2.6
threadpoolctl==3.1.0
torch==2.0.0+cu117
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1+cu117
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
yarl==1.9.2
